apps: move volume operations tests to a new file
